packages/duecare-llm-models/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "duecare-llm-models"
version = "0.1.0"
description = "Duecare model adapters: HF Transformers, llama.cpp, Unsloth, Ollama, OpenAI, Anthropic, Google Gemini, HF Inference Endpoints."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [{name = "Taylor Amarel"}]
keywords = ["llm", "safety", "gemma", "duecare", "agentic"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "duecare-llm-core>=0.1.0,<0.2.0",
]

[project.optional-dependencies]
transformers = [
    "transformers>=4.50.0",
    "accelerate>=1.2.0",
    "bitsandbytes>=0.46.0",
    "torch>=2.4.0",
    "sentencepiece>=0.2.0",
    "tokenizers>=0.21.0",
]
unsloth = [
    "unsloth>=2025.10",
    "peft>=0.15.0",
    "trl>=0.15.0",
    "datasets>=3.0.0",
]
llama-cpp = [
    "llama-cpp-python>=0.3.0",
]
ollama = [
    "ollama>=0.4.0",
]
openai = [
    "openai>=1.50.0",
]
anthropic = [
    "anthropic>=0.40.0",
]
google = [
    "google-generativeai>=0.8.0",
]
hf-endpoint = [
    "huggingface_hub>=0.27.0",
]
all = [
    "duecare-llm-models[transformers]",
    "duecare-llm-models[unsloth]",
    "duecare-llm-models[llama-cpp]",
    "duecare-llm-models[ollama]",
    "duecare-llm-models[openai]",
    "duecare-llm-models[anthropic]",
    "duecare-llm-models[google]",
    "duecare-llm-models[hf-endpoint]",
]



[project.urls]
Homepage = "https://github.com/taylorsamarel/gemma4_comp"
Documentation = "https://github.com/taylorsamarel/gemma4_comp/blob/main/docs/the_forge.md"
Issues = "https://github.com/taylorsamarel/gemma4_comp/issues"
Repository = "https://github.com/taylorsamarel/gemma4_comp"

[tool.hatch.build.targets.wheel]
packages = ["src/duecare"]

[tool.hatch.build.targets.wheel.shared-data]
"PURPOSE.md" = "PURPOSE.md"
"AGENTS.md" = "AGENTS.md"

[tool.hatch.build.targets.sdist]
include = [
    "src/**/*.py",
    "src/**/*.md",
    "src/**/*.yaml",
    "src/**/*.yml",
    "src/**/*.json",
    "tests/**/*.py",
    "README.md",
    "PURPOSE.md",
    "AGENTS.md",
    "pyproject.toml",
]
